# kint2pp: reduce input latency by ≈10ms