--- a/keyboards/kinesis/kint2pp/config.h
+++ b/keyboards/kinesis/kint2pp/config.h
@@ -36,3 +36,7 @@
 #define DEBOUNCE 5
 
 #define IGNORE_MOD_TAP_INTERRUPT
+
+// Reduce input latency by lowering the USB polling interval
+// from its 10ms default to the 1ms minimum that USB 1.x (Full Speed) allows:
+#define USB_POLLING_INTERVAL_MS 1