--- a/keyboards/kinesis/kint2pp/rules.mk
+++ b/keyboards/kinesis/kint2pp/rules.mk
@@ -1,0 +1,3 @@
+# Debounce eagerly (report change immediately), keep per-key timers. We can use
+# this because the kinT does not have to deal with noise.
+DEBOUNCE_TYPE = sym_eager_pk